add yabai and skhd

+++ install.py
@@ -10,12 +10,13 @@ if not os.environ.get('SHELL', '').endswith('zsh'):
     os.exit(1)
 
 if __name__ == "__main__":
-    for fn in ["zshrc", "vimrc", "ideavimrc"]:
+    for fn in ["zshrc", "vimrc", "ideavimrc", "yabairc", "skhdrc"]:
         src = os.path.join(os.getcwd(), fn)
         dst = os.path.join(home, "." + fn) 
         if os.path.exists(dst) and not os.path.islink(dst):
             print("Destination " + dst + " is an existing regular file! Skipping...")
         print("Symlinking from " + src + " to " + dst)
-        os.unlink(dst)
+        if os.path.exists(dst):
+            os.unlink(dst)
         os.symlink(src, dst)
